+++ .eslintrc.cjs
@@ -89,7 +89,13 @@ module.exports = {
         }
       }
     ],
-    'import/no-duplicates': 2
+    'import/no-duplicates': 2,
+    'prefer-arrow/prefer-arrow-functions': [
+      'error',
+      {
+        allowStandaloneDeclarations: true
+      }
+    ]
   },
   overrides: [
     {
@@ -130,6 +136,7 @@ module.exports = {
         '@typescript-eslint/no-unsafe-assignment': 'off',
         '@typescript-eslint/no-unsafe-return': 'off',
         '@typescript-eslint/no-unsafe-call': 'off',
+        '@typescript-eslint/no-floating-promises': 'off',
         '@typescript-eslint/prefer-for-of': 'error',
         '@typescript-eslint/prefer-function-type': 'error',
         '@typescript-eslint/quotes': 'off',