+++ packages/base/src/hooks/index.ts
@@ -1,13 +1,14 @@
 import { useConsolidatedRef } from './useConsolidatedRef';
-import { useI18nBundle } from './useI18nBundle';
+import { I18nBundle, useI18nBundle } from './useI18nBundle';
 import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect';
 import { useIsRTL } from './useIsRTL';
 import { usePassThroughHtmlProps } from './usePassThroughHtmlProps';
 import { useResponsiveContentPadding } from './useResponsiveContentPadding';
-import { useViewportRange } from './useViewportRange';
 import { useSyncRef } from './useSyncRef';
+import { useViewportRange } from './useViewportRange';
 
 export {
+  I18nBundle,
   useConsolidatedRef,
   useI18nBundle,
   useIsomorphicLayoutEffect,