+++ packages/main/src/components/ActionSheet/ActionSheet.jss.ts
@@ -5,13 +5,18 @@ const styles = {
    *  is being applied to the encapsulating ul element
    */
   actionSheet: {
-    margin: 0,
-    padding: '0.1875rem 0.375rem',
+    '&::part(content)': {
+      padding: '0.1875rem 0.375rem'
+    },
     '& [ui5-button]': {
       display: 'block'
     },
     '& [ui5-button]:not([data-cancel-btn])::part(button)': {
       justifyContent: 'flex-start'
+    },
+    '&::part(header)': {
+      background: 'transparent',
+      boxShadow: 'none'
     }
   },
   contentMobile: {