+++ packages/main/src/webComponents/SplitButton/index.tsx
@@ -8,7 +8,7 @@ import { withWebComponent } from '../../internal/withWebComponent';
 
 interface SplitButtonAttributes {
   /**
-   * Sets the accessible aria name of the component.
+   * Defines the accessible aria name of the component.
    */
   accessibleName?: string;
   /**